requirements.txt
python-docx
pyinstaller


git init
git add .
git commit -m "Gerador de reembolso automático"
git branch -M main
git remote add origin https://github.com/unopla/reembolso
git push -u origin main
